requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2>=2.9.10,<3.0
drf-spectacular>=0.27.0,<0.29
Pillow>=11.2.1,<12.0